minor cleanup,
